Add bst.lean
This is mainly to back it up. I tried to make my definitions
as close to those in TaBSTiTL, but that has proven quite
complicated.

Having defined families of expressions using
arbitrary subsets of Variable = Nat, implementing union
of an indexed family of families of expressions would
require knowing which families have which indices, which
is uncomputable (I guess). I'm barely able to work with
what I learned in the zeroth seven chapters of Lean 4
tutorial, and I've had enough. I'm ditching the distinction
between families of expressions and definition lists,
and making the latter indexed by a finite section of
naturals.

As I made this choice, I realized that the advantage
of having infinite definition lists is probably not
as large as I imagined. Since they are required to be
unions of finite FamExprs anyway, and this is an important
aspect of them, plenty of their properties would have
to be considered first in the finite case, or at least I
now think so.

Whatever. If I had someone I could bother with my
Lean-related questions every 5 minutes, I guess I would
push for 1-1 correspondence with TaBSTiTL for (a little)
longer. But one's got more practical issues to solve.