Update etst.pdf and main.text
These are updated from the latest
version in Overleaf. I plan to delete
these documents because I don't want
to keep supporting both a lean
implementation and a latex version,
and also because the latex version
is very outdated anyway. I still want
to preserve what I have in git history,
though. Here it is!