An attempt to remove functions from the definition of Expr
could work in principle, but I'm tired of all this additional work I have
to do to maintain the generality of
Expr. I'll just get rid of signatures.

I'm not happy about this development
in the slightest, but I have to make
practical cuts.

FTR, this would solve a great deal of my
problems:
https://leanprover.zulipchat.com/#narrow/channel/270676-lean4/topic/RFC.3A.20pattern.20matching.20on.20.28syntactically.20equal.29.20functions/with/555144654