Murana Volsi
To be honest, I'm not entirely sure this is a great idea.

The main goal was to simplify Expr by making un,
condSome and arbUn derived concepts (using de
Morgan identities) rather than constructors. The main
expected benefit is a significant simplification of the
proof systems -- rather than having to deal with (mutual)
inductives for both fullness and inhabitedness ^0, and
their corresponding soundness invariants, I hope the rules
for condSome will just "fall out of" the rules for the minimal
system. (Would be funny if I was totally wrong here.)

However, there is also a downside in that the interpretation
function needs to use context for all variables under even
number of complements, else the de Morgan identities
would not hold. As a result, the behavior of the function
is more complicated -- eg. showing monotonicity in context
requires also showing antimonotonicity in background,
whereas before complements were constant in context.
The need to treat even- and odd-times complemented
expressions as separate cases cascades down and increases
complexity in multiple places.

As an aside, this commit also contains some aesthetic
upgrades like Set3.Lane in Chapter 0 instead of Chapter 2.

Notes:
0. See the deleted IsSomeStx in this commit.